conseguindo passar o id do login corretamente, conseguindo chamar o metodo de lista de tarefas, porem algo no layout nao permite a listagem

+++ app/src/main/java/com/example/projetotcc/servico/data/repository/ServicoService.kt
@@ -37,7 +37,7 @@ interface ServicoService {
 
     @GET("api/v1/servico/usuario/{idUsuario}")
     suspend fun buscarTarefasPorIdUsuario(
-        @Path("idUsuario") idUsuario: Long,
+        @Path("idUsuario") idUsuario: Int,
     ): Response<List<ServicoResponse>>
 
     @GET("api/v1/servico/cliente/{idCliente}")
